remove all the components folder refactor
