+++ templates/skeleton/app/components/Footer.tsx
