+++ templates/skeleton/app/components/PaginatedResourceSection.tsx
