+++ templates/skeleton/app/components/ProductImage.tsx
