+++ templates/skeleton/app/components/ProductPrice.tsx
