--- templates/skeleton/app/components/layout/index.ts
@@ -1,1 +1,0 @@
-export {PageLayout} from './PageLayout';
