--- templates/skeleton/app/components/sections/README.md
@@ -1,5 +1,0 @@
-A section in Hydrogen is a reusable, customizable component that contains a visual grouping of components or a user experience flow and can be included in templates or routes.
-
-We recommend doing this so a user experience will be consistent throughout the application and create familiarity to the buyer.
-
-Feel free to modify components in this folder to achieve your desire UX and reuse it throughout your application to create a consistent experience for your buyer.
