--- templates/skeleton/app/components/templates/CartMain/index.ts
@@ -1,1 +1,0 @@
-export {CartMain} from './CartMain';
