--- templates/skeleton/app/components/templates/README.md
@@ -1,3 +1,0 @@
-A template in Hydrogen is a reusable component that define the structure and layout of a route.
-
-This type of components define the structure and layout of a route and is repeated using the data pass in through props.
