+++ templates/skeleton/app/routes/api.predictive-search.tsx
@@ -2,8 +2,8 @@ import {json, type LoaderFunctionArgs} from '@shopify/remix-oxygen';
 import type {
   NormalizedPredictiveSearch,
   NormalizedPredictiveSearchResults,
-} from '~/components/sections/Search';
-import {NO_PREDICTIVE_SEARCH_RESULTS} from '~/components/sections/Search';
+} from '~/components/Search';
+import {NO_PREDICTIVE_SEARCH_RESULTS} from '~/components/Search';
 import {applyTrackingParams} from '~/lib/search';
 
 import type {
